Fix CI: ruff format and react-refresh lint
- Apply ruff format to src/cli.py and tests/test_cli_daemon.py (these
  were untouched on the PR branch but CI's merge with master surfaced
  pre-existing formatting drift)
- Split ThemeContext to satisfy react-refresh/only-export-components:
  move ThemeContext object + useTheme hook to src/hooks/useTheme.ts,
  keep ThemeProvider component alone in src/contexts/ThemeContext.tsx,
  and route consumer imports back through @/hooks/useTheme
