+++ src/frontend/src/hooks/useTheme.ts
@@ -1,3 +1,5 @@
+import { createContext, useContext } from 'react'
+
 export type Theme = 'light' | 'dark'
 
 const STORAGE_KEY = 'mcf:theme'
@@ -25,4 +27,18 @@ function prefersDark(): boolean {
 
 export function resolveInitialTheme(): Theme {
   return readStored() ?? (prefersDark() ? 'dark' : 'light')
+}
+
+export interface ThemeContextValue {
+  theme: Theme
+  setTheme: (theme: Theme) => void
+  toggle: () => void
+}
+
+export const ThemeContext = createContext<ThemeContextValue | null>(null)
+
+export function useTheme(): ThemeContextValue {
+  const ctx = useContext(ThemeContext)
+  if (!ctx) throw new Error('useTheme must be used inside ThemeProvider')
+  return ctx
 }
